convert user/canteen table to listviews
